Make PDF regex pattern work with common name variants
Before this commit the PDF converter script expected the ebook files to be
composed of four strings separated by dashes as in this example:

GLOSSIKA-ENKR-F1-EBK.pdf

This commit addresses the issue raised in #14 and #13 where some users have
PDFs that omit the "-F1" / "-F2" etc from the file name. This portion of the
file pattern is now optional so the following name variants will both work:

GLOSSIKA-ENKR-F1-EBK.pdf
GLOSSIKA-ENKR-EBK.pdf

Note that "GLOSSIKA-" and "-EBK.pdf" are still required to begin and end the
file name respectively. A string beginning with "EN" must immediately follow
"GLOSSIKA-" and is used to identify the languages in the ebook.